feat(config): make quality fractional so a tier can sit between two others
`quality: 1.5` used to truncate to 1 and validate clean. No error, exit 0 — the
model quietly tied the tier it was meant to beat, and nothing said a word. That
silence is the actual bug; a hard parse failure would have been fine.

The motivating case is an MLX 4-bit port of a 27B on a second host: better than
the ternary 27B at quality 1, worse than the original at 2. There is no integer
between them, and renumbering the whole ladder to make room only defers the
problem to the next model that lands between two tiers.

Quality and QualityFloor become float64 through config, lane ordering, the
degrade gate, /v1/models and the dashboard API. Integer configs are unaffected:
1 parses as 1.0 and every comparison behaves as before.

orderCandidates now keys its tier map by float. Exact float equality is the
right test here — a key comes from the same parsed config value every time,
never from arithmetic, so two models written `quality: 1.5` produce
bit-identical keys.

API BREAK: `quality` and `qualityFloor` are Float rather than Int in the
GraphQL schema and JSON. A client decoding them into an integer type needs
updating.

+++ internal/config/config.go
@@ -349,7 +349,17 @@ type Model struct {
 	Swap     *Swap             `yaml:"swap,omitempty"`     // measured load cost (cmd only)
 	Proxy    yaml.Node         `yaml:"proxy,omitempty"`    // forward target: number | "host:port" | {host,port,headers}
 	Type     string            `yaml:"type,omitempty"`     // cost class: chat | embed | openrouter | …
-	Quality  int               `yaml:"quality,omitempty"`
+	// Quality is the relative rank used for lane ordering and degrade gating.
+	// FLOAT, not int: a tier often has to be slotted BETWEEN two existing ones —
+	// an MLX 4-bit port of a model that already sits at 2 is better than the 27B
+	// at 1 and worse than the original, and "1.5" is the only honest way to say
+	// so without renumbering the whole ladder. Integers keep working unchanged;
+	// 1 parses as 1.0.
+	//
+	// Before this was a float, yaml silently TRUNCATED `quality: 1.5` to 1 and
+	// validated clean, so the model quietly tied the tier below the one it was
+	// meant to beat.
+	Quality  float64           `yaml:"quality,omitempty"`
 	// MaxConcurrent is the model's admission slots (the fairshare capacity
 	// unit). For a local llama-server this mirrors --parallel. Default 1.
 	MaxConcurrent int `yaml:"maxConcurrent,omitempty"`
@@ -649,8 +659,8 @@ type Sticky struct {
 
 // MaxQuality returns the highest Quality among the candidates (0 if none/unset)
 // — the top of a served name's quality ladder (P7).
-func MaxQuality(cands []Candidate) int {
-	top := 0
+func MaxQuality(cands []Candidate) float64 {
+	top := 0.0
 	for _, c := range cands {
 		if c.Model.Quality > top {
 			top = c.Model.Quality
@@ -713,7 +723,7 @@ type PriorityGroup struct {
 	AcceptDegrade bool `yaml:"acceptDegrade,omitempty"`
 	// QualityFloor is the lowest backend quality the group will accept when it
 	// does degrade (0 = no floor). Ignored unless AcceptDegrade is set.
-	QualityFloor int `yaml:"qualityFloor,omitempty"`
+	QualityFloor float64 `yaml:"qualityFloor,omitempty"`
 	// PreferResident makes the group best-effort against what is already loaded:
 	// among the backends it accepts (quality-wise), any that are currently
 	// resident (a warm process) are tried first, in quality order, before any
@@ -728,7 +738,7 @@ type PriorityGroup struct {
 // AcceptsQuality reports whether the group may be served by a backend of quality
 // q, given the model's top-tier quality. A non-degrading group accepts only the
 // top tier; a degrading group accepts down to its QualityFloor (P7).
-func (g PriorityGroup) AcceptsQuality(q, topQuality int) bool {
+func (g PriorityGroup) AcceptsQuality(q, topQuality float64) bool {
 	// A model's own top tier is always acceptable — the floor only gates
 	// degrading BELOW the best when a better tier exists. Without this, a group
 	// with QualityFloor>0 rejects any model whose whole ladder sits under the
